requirements.txt
line-bot-sdk==3.5.0
flask==2.3.3
numpy==1.24.3
python-dotenv==1.0.0
gunicorn==21.2.0 